Move common to crate root
