+++ src/core/client/raw.rs
@@ -26,7 +26,7 @@
 
 //! Traits for implementing request-making capabilities.
 
-use crate::core::common;
+use crate::common;
 
 use alloc::boxed::Box;
 use core::{fmt, pin::Pin};
