+++ src/core/server/notification.rs
@@ -24,7 +24,7 @@
 // IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use crate::core::common;
+use crate::common;
 use crate::core::server::Params;
 use core::fmt;
