+++ src/core/server/raw/traits.rs
@@ -24,7 +24,7 @@
 // IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use crate::core::common;
+use crate::common;
 
 use alloc::boxed::Box;
 use core::{hash::Hash, pin::Pin};
